+++ tests/test_result_frame_aggregator.py
@@ -374,3 +374,16 @@ class TestResultFrameAggregator:
         assert list(df_agg_std.values) == pytest.approx(
             [32.83624888, 1.28741954, 3.25905542, 0.4820685]
         )
+
+    @pytest.mark.parametrize("column_mode", ["all", "compact"])
+    def test_lambda_aggregation_override_name(self, res1d_river_network, column_mode):
+        df_reaches = res1d_river_network.reaches.read(column_mode=column_mode)
+        rfa = ResultFrameAggregator(lambda x: x.abs().max(), override_name="max")
+        df_agg = rfa.aggregate(df_reaches)
+
+        assert list(df_agg.columns.values) == [
+            "max_Discharge",
+            "max_FlowVelocity",
+            "max_ManningResistanceNumber",
+            "max_WaterLevel",
+        ]